# src/lib/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Hindus R Us",
    "description": "An API documentation generated by swagger-autogen."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://h-dummy-app-express-app-182702284982.asia-south1.run.app/"
    }
  ],
  "paths": {
...
    "/v1/content/create": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Create Content",
        "description": "An API route to create content - Uploads a file, stores it in Firebase Storage, and creates a content record in Firestore.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentBody"
              }
            }
          }
        }
      }
    },
    "/v1/content/update": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Update Content",
        "description": "An API route to update content.",
        "parameters": [
          {
            "name": "contentid",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Some description...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentBody"
              }
            }
          }
        }
      }
    },
    "/v1/content/update/image": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Update Image",
        "description": "An API route to add a new image in Firebase Storage, and change the thumbnailUrlLink in content collection.",
        "parameters": [
          {
            "name": "contentId",
            "in": "query",
            "description": "The contentId of the content being updated",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Some description...",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateImageBody"
              }
            }
          }
        }
      }
    },
    "/v1/content/data": {
      "get": {
        "tags": [
          "Content"
        ],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Type of content to filter",
            "required": true,
            "example": "educational",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hidden",
            "in": "query",
            "description": "Whether hidden content should be included",
            "required": true,
            "example": "false",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "description": "Language of the content",
            "required": true,
            "example": "english",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateContentBody": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "example": "filename.png"
          },
          "title": {
            "type": "string",
            "example": "Test Content"
          },
          "description": {
            "type": "string",
            "example": "A description written for content creation"
          },
          "language": {
            "type": "string",
            "example": "english"
          },
          "type": {
            "type": "string",
            "example": "educational"
          },
          "hidden": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "file",
          "title",
          "description",
          "language",
          "type"
        ]
      },
      "UpdateContentBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Test Content"
          },
          "description": {
            "type": "string",
            "example": "A description written for content creation"
          },
          "language": {
            "type": "string",
            "example": "english"
          },
          "hidden": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "UpdateImageBody": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "example": "filename.png"
          }
        },
        "required": [
          "file"
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}